Remove some dtype. Still WIP 2

+++ sandbox/crossModalityHamiltonianATSCalibrated_Boundary.py
@@ -92,23 +92,22 @@ def GaussKernelSpaceSingle(sig, d):
 # get A,tau,alpha
 def getATauAlpha(px, qx, pw, qw, cA=1.0, cT=1.0, dimEff=3, single=False):
     xc = (qw * qx).sum(dim=0) / (qw.sum(dim=0))  # moving barycenters; should be 1 x 3
-    A = ((1.0 / (2.0 * cA)) * (px.T @ (qx - xc) - (qx - xc).T @ px)).type(
-        dtype
-    )  # 3 x N * N x 3
-    tau = ((1.0 / cT) * (px.sum(dim=0))).type(dtype)
+    A = ((1.0 / (2.0 * cA)) * (px.T @ (qx - xc) - (qx - xc).T @ px))  # 3 x N * N x 3
+    tau = ((1.0 / cT) * (px.sum(dim=0)))
+
     if dimEff == 2:
-        alpha = ((px * (qx - xc)).sum() + (pw * qw * dimEff).sum()).type(dtype)
-        Alpha = torch.eye(3).type(dtype) * alpha
+        alpha = ((px * (qx - xc)).sum() + (pw * qw * dimEff).sum())
+        Alpha = torch.eye(3) * alpha
         Alpha[-1, -1] = 0.0  # always scale Z by 0
     elif dimEff == 3 and single:
-        alpha = ((px * (qx - xc)).sum() + (pw * qw * dimEff).sum()).type(dtype)
-        Alpha = torch.eye(3).type(dtype) * alpha
+        alpha = ((px * (qx - xc)).sum() + (pw * qw * dimEff).sum())
+        Alpha = torch.eye(3) * alpha
     else:
         alpha_xy = 0.5 * (
             (px[:, 0:2] * (qx - xc)[:, 0:2]).sum() + (pw * qw * 2.0).sum()
-        ).type(dtype)
-        alpha_z = ((px[:, -1] * (qx - xc)[:, -1]).sum() + (pw * qw).sum()).type(dtype)
-        Alpha = torch.eye(3).type(dtype) * alpha_xy
+        )
+        alpha_z = ((px[:, -1] * (qx - xc)[:, -1]).sum() + (pw * qw).sum())
+        Alpha = torch.eye(3) * alpha_xy
         Alpha[-1, -1] = alpha_z
 
     return A, tau, Alpha
@@ -133,7 +132,6 @@ def getU(sigma, d, uCoeff):
 
 def getUdiv(sigma, d, uCoeff):
     xO, qyO, py, wpyO = Vi(0, d), Vj(1, d), Vj(2, d), Vj(3, 1)
-    # retVal = xO.sqdist(qyO)*torch.tensor(0).type(dtype)
     for sInd in range(len(sigma)):
         sig = sigma[sInd]
         x, qy, wpy = xO / sig, qyO / sig, wpyO / sig
